adding session to products

+++ src/routes/products.js
@@ -14,13 +14,7 @@ router.post("/product", bearerAuth, acl("add"), addProduct);
 //get a specific product
 router.get("/product/:id", bearerAuth, acl("read"), getProduct);
 //put
-router.put(
-  "/product/:id",
-  bearerAuth,
-  acl("update"),
-  checkQuantity,
-  updateProduct
-);
+router.put( "/product/:id",bearerAuth,acl("update"), checkQuantity,updateProduct);
 //delete
 router.delete("/product/:id", bearerAuth, acl("remove"), deleteProduct);
 //get one store with the products associated with that store
@@ -40,20 +34,34 @@ async function addProduct(req, res) {
 //gete data of one type of product
 async function getProduct(req, res) {
   const id = req.params.id;
-  res.status(200).json(await products.findOne({ where: { id: id } }));
+  const found=await products.findOne({where: { id: id }})
+  if (found.storeID === req.session.storeID) {
+    res.status(200).json(found);
+  } else {
+    res.status(403).send("Unauthorized access");
+  }
 }
-
+ 
 //update product's data
 async function updateProduct(req, res) {
   const id = req.params.id;
-  const reqBody = req.body;
-  res.status(201).json(await products.update(reqBody, { where: { id: id } }));
+  const oldProduct=await products.findOne({ where: { id: id } });
+  if (oldProduct.storeID === req.session.storeID) {
+    const reqBody = req.body;
+    reqBody.storeID = req.session.storeID;
+    res.status(201).json(await products.update(reqBody, { where: { id: id } }));
+  } else {
+    res.status(403).send("Unauthorized access");
+  }
 }
-
+  
 //delete a product
 async function deleteProduct(req, res) {
   const id = req.params.id;
+  const deletedProduct =await products.findOne({where: { id: id }})
+  if(deleteProduct.storeID===req.session.storeID){
   res.status(200).json(await products.destroy({ where: { id: id } }));
+  }
 }
 
 //get one store products
